+++ trunk/cpsycle/ui/src/uibitmap.h
@@ -30,7 +30,7 @@ struct psy_ui_Bitmap;
 
 typedef void (*psy_ui_bitmap_fp_dispose)(struct psy_ui_Bitmap*);
 typedef int (*psy_ui_bitmap_fp_load)(struct psy_ui_Bitmap*, const char* path);
-typedef int (*psy_ui_bitmap_fp_loadresource)(struct psy_ui_Bitmap*, int resourceid);
+typedef int (*psy_ui_bitmap_fp_loadresource)(struct psy_ui_Bitmap*, uintptr_t resourceid);
 typedef psy_ui_RealSize (*psy_ui_bitmap_fp_size)(const struct psy_ui_Bitmap*);
 typedef bool (*psy_ui_bitmap_fp_empty)(const struct psy_ui_Bitmap*);
 typedef void (*psy_ui_bitmap_fp_settransparency)(struct psy_ui_Bitmap*,
@@ -68,7 +68,7 @@ INLINE int psy_ui_bitmap_load(psy_ui_Bitmap* self, const char* path)
 	return self->vtable->load(self, path);
 }
 
-INLINE int psy_ui_bitmap_loadresource(psy_ui_Bitmap* self, int resourceid)
+INLINE int psy_ui_bitmap_loadresource(psy_ui_Bitmap* self, uintptr_t resourceid)
 {
 	return self->vtable->loadresource(self, resourceid);
 }
@@ -91,7 +91,7 @@ INLINE void psy_ui_bitmap_settransparency(psy_ui_Bitmap* self, psy_ui_Colour col
 // psy_ui_BitmapImp	
 typedef void (*psy_ui_bitmap_imp_fp_dispose)(struct psy_ui_BitmapImp*);
 typedef int (*psy_ui_bitmap_imp_fp_load)(struct psy_ui_BitmapImp*, const char* path);
-typedef int (*psy_ui_bitmap_imp_fp_loadresource)(struct psy_ui_BitmapImp*, int resourceid);
+typedef int (*psy_ui_bitmap_imp_fp_loadresource)(struct psy_ui_BitmapImp*, uintptr_t resourceid);
 typedef psy_ui_RealSize (*psy_ui_bitmap_imp_fp_size)(const struct psy_ui_BitmapImp*);
 typedef bool (*psy_ui_bitmap_imp_fp_empty)(const struct psy_ui_BitmapImp*);
 typedef void (*psy_ui_bitmap_imp_fp_settransparency)(struct psy_ui_BitmapImp*,
